feat: extracted explorer state manager
